fix: Adds CTID to RPC tx and updates error (#4738)
This change fixes a number of issues involved with CTID:
* CTID is not present on all RPC tx transactions.
* rpcWRONG_NETWORK is missing in the ErrorCodes.cpp

+++ src/xrpld/app/ledger/TransactionMaster.h
@@ -80,7 +80,7 @@ public:
         uint256 const& hash,
         std::uint32_t ledger,
         std::optional<uint32_t> tseq,
-        std::optional<uint16_t> netID);
+        std::optional<uint32_t> netID);
 
     void
     canonicalize(std::shared_ptr<Transaction>* pTransaction);
