+++ src/xrpld/app/misc/Transaction.h
@@ -132,7 +132,7 @@ public:
         TransStatus status,
         std::uint32_t ledgerSeq,
         std::optional<uint32_t> transactionSeq = std::nullopt,
-        std::optional<uint16_t> networkID = std::nullopt);
+        std::optional<uint32_t> networkID = std::nullopt);
 
     void
     setStatus(TransStatus status)
@@ -393,8 +393,7 @@ private:
 
     LedgerIndex mLedgerIndex = 0;
     std::optional<uint32_t> mTxnSeq;
-    std::optional<uint16_t> mNetworkID;
-
+    std::optional<uint32_t> mNetworkID;
     TransStatus mStatus = INVALID;
     TER mResult = temUNCERTAIN;
     bool mApplying = false;
